Arrange test names; Implement EqualityComparerTestsGeneral
