+++ src/collections-generic-equalitycomparers/Collections.Generic.EqualityComparers.Tests/ArrayEqualityComparer/EqualityComparerTestsRef.cs
@@ -3,7 +3,7 @@ using Xunit;
 
 namespace PrimeFuncPack.Collections.Generic.EqualityComparers.Tests.ArrayEqualityComparer;
 
-public sealed class EqualityComparerTestsBase_Ref
+public sealed class EqualityComparerTestsRef
 {
     private readonly ArrayEqualityComparer<string?> comparer
         = ArrayEqualityComparer<string?>.Create(CustomEqualityComparer<string?>.Default);
