+++ src/collections-generic-equalitycomparers/Collections.Generic.EqualityComparers.Tests/ImmutableArrayEqualityComparer/EqualityComparerTestsGeneral.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+using System.Collections.Immutable;
+using Xunit;
+
+namespace PrimeFuncPack.Collections.Generic.EqualityComparers.Tests.ImmutableArrayEqualityComparer;
+
+public sealed class EqualityComparerTestsGeneral<T>
+{
+    private readonly ImmutableArrayEqualityComparer<object> comparer
+        = ImmutableArrayEqualityComparer<object>.Create(CustomEqualityComparer<object>.Default);
+
+    [Fact]
+    public void Test_GetHashCode_SourceIsDefault_ExpectZero()
+    {
+        ImmutableArray<object> defaultObj = default;
+        var actual = comparer.GetHashCode(defaultObj);
+        Assert.Equal(0, actual);
+    }
+}
