+++ src/collections-generic-equalitycomparers/Collections.Generic.EqualityComparers.Tests/ListEqualityComparer_List/EqualityComparerTestsRef.cs
@@ -2,16 +2,16 @@
 using System.Linq;
 using Xunit;
 
-namespace PrimeFuncPack.Collections.Generic.EqualityComparers.Tests.ListEqualityComparer.IList;
+namespace PrimeFuncPack.Collections.Generic.EqualityComparers.Tests.ListEqualityComparer_List;
 
-public sealed class EqualityComparerTestsBase_Ref
+public sealed class EqualityComparerTestsRef
 {
     private readonly ListEqualityComparer<string?> comparer
         = ListEqualityComparer<string?>.Create(CustomEqualityComparer<string?>.Default);
 
     [Theory]
     [MemberData(nameof(SourceAreEqualCases))]
-    public void Test_GetHashCode_SourceAreEqual_ExpectHashCodesAreEqual(CaseParamOfIList<string?> source1, CaseParamOfIList<string?> source2)
+    public void Test_GetHashCode_SourceAreEqual_ExpectHashCodesAreEqual(CaseParamOfList<string?> source1, CaseParamOfList<string?> source2)
     {
         var hashCode1 = comparer.GetHashCode(source1.Items);
         var hashCode2 = comparer.GetHashCode(source2.Items);
@@ -20,7 +20,7 @@ public sealed class EqualityComparerTestsBase_Ref
 
     [Theory]
     [MemberData(nameof(SourceAreEqualCases))]
-    public void Test_Equals_SourceAreEqual_ExpectTrue(CaseParamOfIList<string?> source1, CaseParamOfIList<string?> source2)
+    public void Test_Equals_SourceAreEqual_ExpectTrue(CaseParamOfList<string?> source1, CaseParamOfList<string?> source2)
     {
         var actualEquals = comparer.Equals(source1.Items, source2.Items);
         Assert.True(actualEquals);
@@ -28,7 +28,7 @@ public sealed class EqualityComparerTestsBase_Ref
 
     [Theory]
     [MemberData(nameof(SourceAreNotEqualCases))]
-    public void Test_Equals_SourceAreNotEqual_ExpectTrue(CaseParamOfIList<string?> source1, CaseParamOfIList<string?> source2)
+    public void Test_Equals_SourceAreNotEqual_ExpectTrue(CaseParamOfList<string?> source1, CaseParamOfList<string?> source2)
     {
         var actualEquals = comparer.Equals(source1.Items, source2.Items);
         Assert.False(actualEquals);
@@ -37,10 +37,10 @@ public sealed class EqualityComparerTestsBase_Ref
     public static IEnumerable<object[]> SourceAreEqualCases()
         =>
         CaseSourcesArrayRef.SourceAreEqualCases()
-        .Select(@case => CaseMapper.MapToOfIList(@case));
+        .Select(@case => CaseMapper.MapToOfList(@case));
 
     public static IEnumerable<object[]> SourceAreNotEqualCases()
         =>
         CaseSourcesArrayRef.SourceAreNotEqualCases()
-        .Select(@case => CaseMapper.MapToOfIList(@case));
+        .Select(@case => CaseMapper.MapToOfList(@case));
 }
