+++ src/collections-generic-equalitycomparers/Collections.Generic.EqualityComparers.Tests/ReadOnlyListEqualityComparer/EqualityComparerTestsStruct.cs
@@ -2,16 +2,16 @@
 using System.Linq;
 using Xunit;
 
-namespace PrimeFuncPack.Collections.Generic.EqualityComparers.Tests.ListEqualityComparer.List;
+namespace PrimeFuncPack.Collections.Generic.EqualityComparers.Tests.ReadOnlyListEqualityComparer;
 
-public sealed class EqualityComparerTestsBase_Struct
+public sealed class EqualityComparerTestsStruct
 {
-    private readonly ListEqualityComparer<int?> comparer
-        = ListEqualityComparer<int?>.Create(CustomEqualityComparer<int?>.Default);
+    private readonly ReadOnlyListEqualityComparer<int?> comparer
+        = ReadOnlyListEqualityComparer<int?>.Create(CustomEqualityComparer<int?>.Default);
 
     [Theory]
     [MemberData(nameof(SourceAreEqualCases))]
-    public void Test_GetHashCode_SourceAreEqual_ExpectHashCodesAreEqual(CaseParamOfList<int?> source1, CaseParamOfList<int?> source2)
+    public void Test_GetHashCode_SourceAreEqual_ExpectHashCodesAreEqual(CaseParamOfIReadOnlyList<int?> source1, CaseParamOfIReadOnlyList<int?> source2)
     {
         var hashCode1 = comparer.GetHashCode(source1.Items);
         var hashCode2 = comparer.GetHashCode(source2.Items);
@@ -20,7 +20,7 @@ public sealed class EqualityComparerTestsBase_Struct
 
     [Theory]
     [MemberData(nameof(SourceAreEqualCases))]
-    public void Test_Equals_SourceAreEqual_ExpectTrue(CaseParamOfList<int?> source1, CaseParamOfList<int?> source2)
+    public void Test_Equals_SourceAreEqual_ExpectTrue(CaseParamOfIReadOnlyList<int?> source1, CaseParamOfIReadOnlyList<int?> source2)
     {
         var actualEquals = comparer.Equals(source1.Items, source2.Items);
         Assert.True(actualEquals);
@@ -28,7 +28,7 @@ public sealed class EqualityComparerTestsBase_Struct
 
     [Theory]
     [MemberData(nameof(SourceAreNotEqualCases))]
-    public void Test_Equals_SourceAreNotEqual_ExpectTrue(CaseParamOfList<int?> source1, CaseParamOfList<int?> source2)
+    public void Test_Equals_SourceAreNotEqual_ExpectTrue(CaseParamOfIReadOnlyList<int?> source1, CaseParamOfIReadOnlyList<int?> source2)
     {
         var actualEquals = comparer.Equals(source1.Items, source2.Items);
         Assert.False(actualEquals);
@@ -37,10 +37,10 @@ public sealed class EqualityComparerTestsBase_Struct
     public static IEnumerable<object[]> SourceAreEqualCases()
         =>
         CaseSourcesArrayStruct.SourceAreEqualCases()
-        .Select(@case => CaseMapper.MapToOfList(@case));
+        .Select(@case => CaseMapper.MapToOfIReadOnlyList(@case));
 
     public static IEnumerable<object[]> SourceAreNotEqualCases()
         =>
         CaseSourcesArrayStruct.SourceAreNotEqualCases()
-        .Select(@case => CaseMapper.MapToOfList(@case));
+        .Select(@case => CaseMapper.MapToOfIReadOnlyList(@case));
 }
